renamed alpha_banner component
